efficientvit modified to be used as a backbone in attention unet

+++ keras_unet_collection/layer_utils.py
@@ -428,11 +428,11 @@ def CONV_output(X, n_labels, kernel_size=1, activation='Softmax', name='conv_out
     if activation:
         
         if activation == 'Sigmoid':
-            X = Activation('sigmoid', name='{}_activation'.format(name))(X)
+            X = Activation('sigmoid', dtype="float32", name='{}_activation'.format(name))(X)
             
         else:
             activation_func = eval(activation)
-            X = activation_func(name='{}_activation'.format(name))(X)
+            X = activation_func(dtype="float32",name='{}_activation'.format(name))(X)
             
     return X
